+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/Sketch.common.kt
@@ -40,7 +40,7 @@ import com.github.panpf.sketch.request.internal.PlaceholderRequestInterceptor
 import com.github.panpf.sketch.request.internal.RequestExecutor
 import com.github.panpf.sketch.source.ByteArrayDataSource
 import com.github.panpf.sketch.source.FileDataSource
-import com.github.panpf.sketch.transform.internal.TransformationDecodeInterceptor
+import com.github.panpf.sketch.transform.internal.TransformationRequestInterceptor
 import com.github.panpf.sketch.util.ComponentLoader
 import com.github.panpf.sketch.util.DecoderProvider
 import com.github.panpf.sketch.util.DownloadData
@@ -535,8 +535,8 @@ internal fun commonComponents(): ComponentRegistry = ComponentRegistry {
     addRequestInterceptor(MemoryCacheRequestInterceptor())
     addRequestInterceptor(PlaceholderRequestInterceptor())
     addRequestInterceptor(ResultCacheRequestInterceptor())
+    addRequestInterceptor(TransformationRequestInterceptor())
     addRequestInterceptor(EngineRequestInterceptor())
 
-    addDecodeInterceptor(TransformationDecodeInterceptor())
     addDecodeInterceptor(EngineDecodeInterceptor())
 }